refactor(result): split large result.h into modular sub-headers
Split 913-line result.h into focused sub-headers under result/ directory
to improve compilation times and maintainability:

- result/fwd.h: Forward declarations and result_state enum
- result/error_info.h: error_info struct and error_code alias
- result/result_core.h: Result<T> class definition
- result/optional.h: Optional<T> class
- result/result_funcs.h: Factory and helper functions
- result/error_codes_compat.h: Backward compatibility aliases
- result/exception_conversion.h: try_catch utilities
- result/result_macros.h: Convenience macros

result.h becomes an umbrella header that includes all sub-headers,
maintaining full backward compatibility with existing code.

Closes #243
